Organize API into folders
