+++ apps/dashboard/src/routes/+page.svelte
@@ -61,7 +61,7 @@
 			const { subdomain, domain, orgId } = res;
 
 			res.isRequesting(true);
-			const domainRes = await fetch('/api/create-url', {
+			const domainRes = await fetch('/api/url/create', {
 				method: 'POST',
 				body: JSON.stringify({ subdomain: subdomain, domain: domain, orgId: orgId })
 			});
@@ -95,7 +95,7 @@
 			const { name } = res;
 
 			res.isRequesting(true);
-			const orgRes = await fetch('/api/create-org', {
+			const orgRes = await fetch('/api/org/create', {
 				method: 'POST',
 				body: JSON.stringify({ name: name })
 			});
