+++ apps/dashboard/src/routes/api/org/create/+server.ts
