+++ apps/dashboard/src/routes/api/url/delete/+server.ts
